src/requirements.txt
psycopg2-binary>=2.9.0
pandas>=2.0.0
boto3>=1.28.0
aiohttp>=3.9.0
requests>=2.31.0
configparser>=5.3.0